+++ dev-packages/cli-common/src/utils/component-util.ts
@@ -1,6 +1,7 @@
 import { readFileSync, existsSync } from 'fs-extra';
 import * as lockfile from '@yarnpkg/lockfile';
 import * as path from 'path';
+import { MALAGU_COMPONENT_SUFFIX } from '../package/package-protocol';
 const chalk = require('chalk');
 
 export namespace ComponentUtil {
@@ -51,6 +52,13 @@ export namespace ComponentUtil {
                 }
             });
         }
+    }
+
+    export function getComponentAlias(keywords: string[] = []): string[] {
+        const keywordsAlias = keywords
+            .filter(k => k.endsWith(MALAGU_COMPONENT_SUFFIX))
+            .map(k => k.substring(0, k.length - MALAGU_COMPONENT_SUFFIX.length));
+        return keywords.includes('malagu') ? [ ...keywordsAlias, 'malagu' ] : keywordsAlias;
     }
 }