+++ packages/core/src/common/config/component.ts
@@ -1,3 +1,0 @@
-export enum Scope {
-    Request = 1, Singleton = 2, Transient  = 3
-}